refactor ai.agents.planner (core part) to ai.agents.core.planner to avoid name clashes
