+++ agents/agents-features/agents-features-snapshot/src/jvmTest/kotlin/ai/koog/agents/snapshot/PlannerPersistenceTests.kt
@@ -3,10 +3,10 @@ package ai.koog.agents.snapshot
 import ai.koog.agents.core.agent.AIAgent
 import ai.koog.agents.core.agent.config.AIAgentConfig
 import ai.koog.agents.core.agent.context.AIAgentPlannerContext
+import ai.koog.agents.core.planner.AIAgentPlanner
+import ai.koog.agents.core.planner.AIAgentPlannerStrategy
+import ai.koog.agents.core.planner.PlannerAgentExecutionPoint
 import ai.koog.agents.core.tools.ToolDescriptor
-import ai.koog.agents.planner.AIAgentPlanner
-import ai.koog.agents.planner.AIAgentPlannerStrategy
-import ai.koog.agents.planner.PlannerAgentExecutionPoint
 import ai.koog.agents.planner.goap
 import ai.koog.agents.planner.goap.GoapAgentState
 import ai.koog.agents.planner.llm.PlanStep
