+++ agents/agents-planner/src/commonMain/kotlin/ai/koog/agents/planner/llm/SimpleLLMPlanner.kt
@@ -3,7 +3,7 @@ package ai.koog.agents.planner.llm
 import ai.koog.agents.core.agent.context.AIAgentPlannerContext
 import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.dsl.extension.HistoryCompressionStrategy
-import ai.koog.agents.planner.AIAgentPlanner
+import ai.koog.agents.core.planner.AIAgentPlanner
 import ai.koog.prompt.dsl.prompt
 import ai.koog.prompt.markdown.markdown
 import ai.koog.prompt.message.Message
