+++ integration-tests/src/jvmTest/java/ai/koog/integration/tests/agent/JavaPlannerAIAgentIntegrationTest.java
@@ -4,11 +4,9 @@ import ai.koog.agents.core.agent.AIAgent;
 import ai.koog.agents.core.agent.context.AIAgentPlannerContext;
 import ai.koog.agents.core.tools.ToolRegistry;
 import ai.koog.agents.core.tools.ToolRegistryBuilder;
-import ai.koog.agents.planner.AIAgentPlannerStrategy;
+import ai.koog.agents.core.planner.AIAgentPlannerStrategy;
 import ai.koog.agents.planner.JavaAIAgentPlanner;
 import ai.koog.agents.planner.Planners;
-import ai.koog.agents.planner.goap.Action;
-import ai.koog.agents.planner.goap.Goal;
 import ai.koog.agents.planner.goap.GoapAgentState;
 import ai.koog.integration.tests.base.KoogJavaTestBase;
 import ai.koog.integration.tests.utils.NumberTools;
